Fix tests and lints
